Notes support, new generator API
